+++ src/Coypu.AcceptanceTests/ApiExamples.cs
@@ -52,7 +52,12 @@ namespace Coypu.AcceptanceTests
 
         private void ReloadTestPage()
         {
-            browser.Visit("file:///" + new FileInfo(@"html\InteractionTestsPage.htm").FullName.Replace("\\", "/"));
+            browser.Visit(TestPageLocation("InteractionTestsPage.htm"));
+        }
+
+        private static string TestPageLocation(string page)
+        {
+            return "file:///" + new FileInfo(@"html\" + page).FullName.Replace("\\", "/");
         }
 
         private void ReloadTestPageWithDelay()
@@ -90,7 +95,7 @@ namespace Coypu.AcceptanceTests
         public void Uncheck_example()
         {
             browser.Uncheck("checkedBox");
-            Assert.IsFalse(browser.FindField("checkedBox").Selected);
+            Assert.IsFalse(browser.Query(() => browser.FindField("checkedBox").Selected,false));
         }
 
         [Test]
@@ -409,6 +414,20 @@ namespace Coypu.AcceptanceTests
 
             Assert.That(expectingScope1.Id, Is.EqualTo("iframe1ButtonId"));
             Assert.That(expectingScope2.Id, Is.EqualTo("iframe2ButtonId"));
+        }
+
+        [Test]
+        public void WithinFrame_example()
+        {
+            browser.Visit(TestPageLocation("frameset.htm"));
+
+            const string selectorThatAppearsInMultipleScopes = "scoped button";
+
+            var expectingScope1 = browser.FindFrame("frame1").FindButton(selectorThatAppearsInMultipleScopes);
+            var expectingScope2 = browser.FindFrame("frame2").FindButton(selectorThatAppearsInMultipleScopes);
+
+            Assert.That(expectingScope1.Id, Is.EqualTo("frame1ButtonId"));
+            Assert.That(expectingScope2.Id, Is.EqualTo("frame2ButtonId"));
         }
 
         [Test]