+++ src/Coypu.Drivers.Tests/When_finding_iframes.cs
@@ -8,22 +8,22 @@ namespace Coypu.Drivers.Tests
         [Test]
         public void Finds_by_header_text()
         {
-            Driver.FindIFrame("I am iframe one", Root).Id.should_be("iframe1");
-            Driver.FindIFrame("I am iframe two", Root).Id.should_be("iframe2");
+            Driver.FindFrame("I am iframe one", Root).Id.should_be("iframe1");
+            Driver.FindFrame("I am iframe two", Root).Id.should_be("iframe2");
         }
 
         [Test]
         public void Finds_by_id()
         {
-            Driver.FindIFrame("iframe1", Root).Id.should_be("iframe1");
-            Driver.FindIFrame("iframe2", Root).Id.should_be("iframe2");
+            Driver.FindFrame("iframe1", Root).Id.should_be("iframe1");
+            Driver.FindFrame("iframe2", Root).Id.should_be("iframe2");
         }
 
         [Test]
         public void Finds_by_title()
         {
-            Driver.FindIFrame("iframe one title", Root).Id.should_be("iframe1");
-            Driver.FindIFrame("iframe two title", Root).Id.should_be("iframe2");
+            Driver.FindFrame("iframe one title", Root).Id.should_be("iframe1");
+            Driver.FindFrame("iframe two title", Root).Id.should_be("iframe2");
         }
     }
 }