+++ src/Coypu.Tests/When_interacting_with_the_browser/When_inspecting_location.cs
@@ -10,7 +10,7 @@ namespace Coypu.Tests.When_interacting_with_the_browser
         public void It_returns_the_driver_url()
         {
             var driverLocation = new Uri("https://blank.org:8080/actual_location");
-            driver.StubLocation(driverLocation);
+            driver.StubLocation(driverLocation,browserSession);
             Assert.That(browserSession.Location, Is.EqualTo(driverLocation));
         }
     }