+++ src/Coypu/DriverScope.cs
@@ -45,7 +45,7 @@ namespace Coypu
 
         public virtual Uri Location
         {
-            get { return driver.Location; }
+            get { return driver.Location(this); }
         }
 
         public bool ConsiderInvisibleElements
@@ -250,9 +250,15 @@ namespace Coypu
             Query(action);
         }
 
-        public IFrameElementScope FindIFrame(string locator, Options options = null)
+        public ElementScope FindFrame(string locator, Options options = null)
         {
-            return new IFrameElementScope(new IFrameFinder(driver, locator, this), this, SetOptions(options));
+            return new RobustElementScope(new FrameFinder(driver, locator, this), this, SetOptions(options));
+        }
+
+        [Obsolete("Use FindFrame which will find both iframes and old school frameset frames. This method will be removed in version 0.9")]
+        public ElementScope FindIFrame(string locator, Options options = null)
+        {
+            return FindFrame(locator, options);
         }
 
         public T Query<T>(Func<T> query, T expecting, Options options = null)